perf: change double to float

+++ src/main/java/com/finances/api/data/protocols/createTransactionRepository/CreateTransactionRepositoryDto.java
@@ -3,7 +3,7 @@ package com.finances.api.data.protocols.createTransactionRepository;
 import java.util.UUID;
 
 public record CreateTransactionRepositoryDto(
-        Double value,
+        float value,
         String type,
         UUID fromAccountId,
         UUID targetAccountId
